refactor: centralize toast hook
