+++ components/add-to-cart-button.tsx
@@ -3,7 +3,7 @@
 import { Button } from "@/components/ui/button"
 import type { Product } from "@/lib/types"
 import { useCart } from "./cart-provider"
-import { useToast } from "@/components/ui/use-toast"
+import { useToast } from "@/hooks/use-toast"
 import { useState } from "react"
 import { Minus, Plus, ShoppingCart } from "lucide-react"
